% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{is_fastpipe}
\alias{is_fastpipe}
\title{Test if Object is a fastpipe}
\usage{
is_fastpipe(x)
}
\arguments{
\item{x}{object to test}
}
\value{
a length one logical
}
\description{
Test if Object is a fastpipe
}
