% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{pipes}
\alias{pipes}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%T>\%}
\alias{\%$\%}
\title{Pipes}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs

lhs \%T>\% rhs

lhs \%$\% rhs
}
\arguments{
\item{lhs}{A value or a dot (\code{.}).}

\item{rhs}{A function call using pipe semantics of the relevant pipe.}
}
\description{
Equivalents to magrittr's pipes
}
