% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{print.fastpipe}
\alias{print.fastpipe}
\title{Print a fastpipe object}
\usage{
\method{print}{fastpipe}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{Ignored, kept for compatibility with other methods}
}
\description{
Print a fastpipe object
}
